+++ changelog.md
@@ -34,6 +34,7 @@
 
 # Unreleased
 
+- Fix NmsVersion with 1.8 and 1.9
 - Fix MenuItemStack loader with placeholder check
 - Fix CONTAINS_STRING action for placeholder
 - Add matrix support for button slots [#39](https://github.com/Maxlego08/zMenu/issues/39)